Need to fix segfault

+++ core/src/anchor.h
@@ -9,7 +9,7 @@ typedef struct Anchor {
     int last_anchor_col;
     int transpose_state;
     int vertical;
-    double highest_possible_equity;
+    float highest_possible_equity;
 } Anchor;
 
 typedef struct AnchorList {
@@ -19,7 +19,7 @@ typedef struct AnchorList {
 
 AnchorList * create_anchor_list();
 void destroy_anchor_list(AnchorList * al);
-void insert_anchor(AnchorList* al, int row, int col, int last_anchor_col, int transpose_state, int vertical, double highest_possible_equity);
+void insert_anchor(AnchorList* al, int row, int col, int last_anchor_col, int transpose_state, int vertical, float highest_possible_equity);
 void reset_anchor_list(AnchorList * al);
 
 #endif
